GUACAMOLE-1218: Refactor guacamole-auth-json's ByteArrayProperty into guacamole-ext.
